crypto_bot/requirements.txt
requests
python-binance
numpy
flask